+++ tests/unit/behaviors/EmailTemplateBehaviorTest.php
@@ -11,6 +11,8 @@ use yii\helpers\Json;
 use ymaker\email\templates\behaviors\EmailTemplateBehavior;
 use ymaker\email\templates\entities\EmailTemplate;
 use ymaker\email\templates\entities\EmailTemplateTranslation;
+use ymaker\email\templates\repositories\EmailTemplatesRepository;
+use ymaker\email\templates\repositories\EmailTemplatesRepositoryInterface;
 use ymaker\email\templates\tests\mocks\DemoActiveRecord;
 use ymaker\email\templates\tests\unit\TestCase;
 
@@ -24,14 +26,25 @@ use ymaker\email\templates\tests\unit\TestCase;
  */
 class EmailTemplateBehaviorTest extends TestCase
 {
+    /**
+     * @inheritdoc
+     */
+    protected function _inject()
+    {
+        \Yii::$container->set(
+            EmailTemplatesRepositoryInterface::class,
+            EmailTemplatesRepository::class
+        );
+    }
+
     public function testNewModel()
     {
         DemoActiveRecord::$behaviors = [EmailTemplateBehavior::class];
-        $model = new DemoActiveRecord([
-            'letterSubject' => 'this is subject',
-            'letterBody' => 'this is body',
-            'emailTemplateHint' => 'this is hint',
-        ]);
+
+        $model = new DemoActiveRecord();
+        $model->letterSubject = 'this is subject';
+        $model->letterBody = 'this is body';
+        $model->emailTemplateHint = 'this is hint';
         $model->save(false);
 
         $this->tester->seeRecord(EmailTemplate::class, [
@@ -51,11 +64,11 @@ class EmailTemplateBehaviorTest extends TestCase
     public function testFindModel()
     {
         DemoActiveRecord::$behaviors = ['templates' => EmailTemplateBehavior::class];
-        $model = new DemoActiveRecord([
-            'letterSubject' => 'this is subject',
-            'letterBody' => 'this is body',
-            'emailTemplateHint' => 'this is hint',
-        ]);
+
+        $model = new DemoActiveRecord();
+        $model->letterSubject = 'this is subject';
+        $model->letterBody = 'this is body';
+        $model->emailTemplateHint = 'this is hint';
         $model->save(false);
 
         $foundedModel = DemoActiveRecord::findOne($model->id);